Move External Intent tests
