+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/externalintents/ExportedActivityTestRule.java
@@ -1,9 +1,9 @@
-package org.odk.collect.android.externalintents;
+package org.odk.collect.android.feature.externalintents;
 
 import android.app.Activity;
 import androidx.test.rule.ActivityTestRule;
 
-import static org.odk.collect.android.externalintents.ExportedActivitiesUtils.clearDirectories;
+import static org.odk.collect.android.feature.externalintents.ExportedActivitiesUtils.clearDirectories;
 
 class ExportedActivityTestRule<A extends Activity> extends ActivityTestRule<A> {
 
@@ -14,7 +14,6 @@ class ExportedActivityTestRule<A extends Activity> extends ActivityTestRule<A> {
     @Override
     protected void beforeActivityLaunched() {
         super.beforeActivityLaunched();
-
         clearDirectories();
     }
