+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/externalintents/FormEntryActivityTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.externalintents;
+package org.odk.collect.android.feature.externalintents;
 
 import androidx.test.filters.Suppress;
 import androidx.test.rule.ActivityTestRule;
@@ -11,7 +11,7 @@ import org.odk.collect.android.activities.FormEntryActivity;
 
 import java.io.IOException;
 
-import static org.odk.collect.android.externalintents.ExportedActivitiesUtils.testDirectories;
+import static org.odk.collect.android.feature.externalintents.ExportedActivitiesUtils.testDirectories;
 
 @Suppress
 // Frequent failures: https://github.com/opendatakit/collect/issues/796
